refactor(indexing): extract snippet repository and mapper
